Create abstract go codegen base class
Remove tabs
