+++ samples/server/petstore/go-api-server/go/README.md
@@ -13,7 +13,7 @@ To see how to make this your own, look here:
 [README](https://github.com/swagger-api/swagger-codegen/blob/master/README.md)
 
 - API version: 1.0.0
-- Build date: 2017-12-17T09:38:04.917Z
+- Build date: 2017-12-18T09:35:47.160Z
 
 
 ### Running the server
